# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ZELO — API de cuidado compartilhado
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /families/{familyId}/caregivers:
    get:
      operationId: listCaregivers
      tags: [caregivers]
      parameters:
        - { name: familyId, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Caregiver"
    post:
      operationId: createCaregiver
      tags: [caregivers]
      parameters:
        - { name: familyId, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaregiverInput"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Caregiver"
  /families/{familyId}/caregivers/{caregiverId}:
    get:
      operationId: getCaregiver
      tags: [caregivers]
      parameters:
        - { name: familyId, in: path, required: true, schema: { type: number } }
        - { name: caregiverId, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Caregiver"
        "404":
          description: Não encontrado
    patch:
      operationId: updateCaregiver
      tags: [caregivers]
      parameters:
        - { name: familyId, in: path, required: true, schema: { type: number } }
        - { name: caregiverId, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaregiverUpdate"
      responses:
        "200":
          description: Atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Caregiver"

  # ── Medicamentos ──────────────────────────────────────────────────────────
...
  /families/{familyId}/audit-log:
    get:
      operationId: listAuditLog
      tags: [audit]
      summary: Log de auditoria imutável — somente leitura
      parameters:
        - { name: familyId, in: path, required: true, schema: { type: number } }
        - name: from
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: entityType
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: number, default: 50, maximum: 200 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"

  # ── Assinaturas ───────────────────────────────────────────────────────────
...
  /families/{familyId}/dashboard:
    get:
      operationId: getFamilyDashboard
      tags: [dashboard]
      summary: Resumo do painel da família
      parameters:
        - { name: familyId, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyDashboard"
  /families/{familyId}/patients/{patientId}/today-doses:
    get:
      operationId: getTodayDoses
      tags: [dashboard]
      summary: Doses de hoje do paciente (fuso do paciente)
      parameters:
        - { name: familyId, in: path, required: true, schema: { type: number } }
        - { name: patientId, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodayDoseSummary"
  /families/{familyId}/patients/{patientId}/adherence-stats:
    get:
      operationId: getAdherenceStats
      tags: [dashboard]
      summary: Estatísticas de adesão por período
      parameters:
        - { name: familyId, in: path, required: true, schema: { type: number } }
        - { name: patientId, in: path, required: true, schema: { type: number } }
        - name: days
          in: query
          required: false
          schema: { type: number, default: 30 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdherenceStats"

  # ── Design Reference ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    Caregiver:
      type: object
      required: [id, familyId, name, role, createdAt]
      properties:
        id:        { type: number }
        familyId:  { type: number }
        name:      { type: string }
        email:     { type: ["string", "null"] }
        role:
          type: string
          enum: [primary_caregiver, caregiver, hired_caregiver, observer]
          description: "primary_caregiver=responsável principal; observer=parente distante que só acompanha"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: ["string", "null"], format: date-time }
    CaregiverInput:
      type: object
      required: [name, role]
      properties:
        name:  { type: string, minLength: 1 }
        email: { type: ["string", "null"] }
        role:
          type: string
          enum: [primary_caregiver, caregiver, hired_caregiver, observer]
    CaregiverUpdate:
      type: object
      properties:
        name:  { type: string }
        email: { type: ["string", "null"] }
        role:
          type: string
          enum: [primary_caregiver, caregiver, hired_caregiver, observer]

    # ── Medication ───────────────────────────────────────────────────────────
...
    ScheduledDose:
      type: object
      required: [id, treatmentId, patientId, scheduledAt, status]
      properties:
        id:           { type: number }
        treatmentId:  { type: number }
        patientId:    { type: number }
        medicationName: { type: string }
        dose:         { type: ["string", "null"] }
        scheduledAt:  { type: string, format: date-time, description: "UTC. Converter para fuso do paciente ao exibir." }
        status:
          type: string
          enum: [pending, taken, skipped, late]
        doseRecord:   { $ref: "#/components/schemas/DoseRecord", nullable: true }

    # ── DoseRecord ───────────────────────────────────────────────────────────
    DoseRecord:
      type: object
      required: [id, scheduledDoseId, patientId, caregiverId, takenAt, createdAt]
      properties:
        id:              { type: number }
        scheduledDoseId: { type: number }
        patientId:       { type: number }
        caregiverId:     { type: number }
        caregiverName:   { type: string }
        takenAt:         { type: string, format: date-time }
        outcome:
          type: string
          enum: [taken, skipped]
        notes:           { type: ["string", "null"] }
        createdAt:       { type: string, format: date-time }
...
    AuditLogEntry:
      type: object
      required: [id, familyId, entityType, entityId, action, actorId, actorType, createdAt]
      description: "Imutável. Nunca pode ser editado nem apagado."
      properties:
        id:          { type: number }
        familyId:    { type: number }
        entityType:  { type: string }
        entityId:    { type: string }
        action:      { type: string, enum: [created, updated, deleted, accessed] }
        actorId:     { type: ["string", "null"] }
        actorType:   { type: string, enum: [caregiver, system] }
        diff:        { type: ["string", "null"], description: "JSON das mudanças (campos não sensíveis)" }
        ipAddress:   { type: ["string", "null"] }
        userAgent:   { type: ["string", "null"] }
        createdAt:   { type: string, format: date-time }

    # ── Subscription ─────────────────────────────────────────────────────────
...
    FamilyDashboard:
      type: object
      required: [familyId, patientCount, caregiverCount, pendingDosesToday, takenDosesToday, lateDoses]
      properties:
        familyId:         { type: number }
        patientCount:     { type: number }
        caregiverCount:   { type: number }
        pendingDosesToday: { type: number }
        takenDosesToday:  { type: number }
        lateDoses:        { type: number }
        upcomingAppointments: { type: number }
        lowStockItems:    { type: number }
    TodayDoseSummary:
      type: object
      required: [date, patientTimezone, totalDoses, takenDoses, pendingDoses, lateDoses, doses]
      properties:
        date:            { type: string, format: date }
        patientTimezone: { type: string }
        totalDoses:      { type: number }
        takenDoses:      { type: number }
        pendingDoses:    { type: number }
        lateDoses:       { type: number }
        doses:
          type: array
          items:
            $ref: "#/components/schemas/ScheduledDose"
    AdherenceStats:
      type: object
      required: [patientId, periodDays, totalScheduled, totalTaken, totalSkipped, totalLate, adherenceRate]
      properties:
        patientId:      { type: number }
        periodDays:     { type: number }
        totalScheduled: { type: number }
        totalTaken:     { type: number }
        totalSkipped:   { type: number }
        totalLate:      { type: number }
        adherenceRate:  { type: number, description: "0.0 a 1.0" }

    # ── Design Tokens ─────────────────────────────────────────────────────────
...